Updated format of technologies section
Technologies sectiion was now split into two sub sections, namely software and hardware, in future, for example, leap technology could be added to the hardware subsection (was listed as a nice to have in the client requirements). Added considered alternatives to both listed choices for the game engine and the VR headset, and listed what the choices provide.